Add 'commit-starts-with' optional parameter
An optional parameter has been added to validate that the commit being
evaluated starts with a specific string.

If this optional parameter is set, we will check the subject line of
the current commit to ensure it matches what was expected. We will
replace the string "[version]" in this parameter with the current
release version.

The README has been updated to explain how this works.