Add GitHub workflows
Currently in order to make a new release of this Action one must
manually create and push up a Git tag. This does not follow the same
process we use for our other Actions, such as `action-publish-release`.

In fact, this repo lacks a `.github` directory entirely. To fix this,
copy this directory from `action-publish-release`. Note that the
`build-lint-test` workflow is different, since we don't have any
JavaScript code in this action; instead we just run `shellcheck`. We
also don't have any configuration for Dependabot for the same reason.