Detect whether a release is being rolled back
If a commit is pushed which rolls back a release candidate and hence
downgrades the version of the package in question, then this action will
regard it as a release commit. This isn't right, and we should do
something about it.

However, we don't simply want to ignore this information, because it
may come in handy (for instance, if we want to automatically remove the
GitHub release/tag or verify the changelog in a different way).

Therefore, this commit adds a new output, `COMMIT_TYPE`, which can
either be `release`, `release-rollback`, or unset (not a release or a
release rollback). In fact, this commit makes this output recommended
and deprecates `IS_RELEASE`.